requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
gspread==5.12.0
google-auth==2.23.4
